ahi anda mejor
va queriendo, ya no confunde las patas.. pero falta ponerle un poco de
deteccion de errores al SSE nuevo y habilitar el seteo de patas
diferenciales con SDI

+++ Source/fuentes/conversor_logic.c
@@ -5,7 +5,7 @@
 int cont = 0;
 short int posicion = 0;
 short int dato_n;
-unsigned int num;
+unsigned short int num;
